requirements.txt
notion-client
google-generativeai
python-dotenv
